requirements.txt
altgraph==0.16.1
amqp==2.5.1
aniso8601==6.0.0
appnope==0.1.0
argcomplete==1.10.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.8.0
billiard==3.6.1.0
bs4==0.0.1
build==1.0.2
cached-property==1.5.1
cachetools==3.1.1
celery==4.3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudconvert==1.0.0
comtypes==1.1.7
config==0.4.2
configparser==3.7.4
constantly==15.1.0
convertapi==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
cs50==3.1.0
cssselect==1.0.3
decorator==4.4.0
defusedxml==0.6.0
dj-static==0.0.6
Django
django-allauth==0.39.1
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
dnspython==1.16.0
docx2txt==0.8
drf-yasg==1.16.1
dss-python==0.10.3
EbookLib==0.17.1
entrypoints==0.3
exchangelib==1.12.5
extract-msg==0.23.1
Flask==1.0.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
Flask-Session==0.3.1
Flask-SocketIO==4.1.0
forex-python==1.5
fpdf==1.7.2
future==0.17.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-calendar-api-client-python==0.1.0
gunicorn==19.9.0
httplib2==0.13.0
hyperlink==18.0.0
icalendar==4.0.3
idna==2.8
IMAPClient==2.1.0
importlib-metadata==0.18
incremental==17.5.0
inflection==0.3.1
ipaddress==1.0.22
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
keyring==19.0.2
kombu==4.6.4
lxml==4.3.3
macholib==1.11
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==2.19.2
mongo==0.2.0
more-itertools==7.0.0
mysql-connector==2.2.9
mysql-connector-python==8.0.16
ntlm-auth==1.3.0
numpy==1.16.2
O365==2.0.1
oauth2client==4.1.3
oauthlib==3.0.1
olefile==0.46
packaging==19.0
pandoc==1.0.2
parsel==1.5.1
parso==0.4.0
pdfminer.six==20181108
pexpect==4.7.0
pickleshare==0.7.5
pika==1.1.0
Pillow==6.0.0
pluggy==0.12.0
ply==3.11
pprint==0.1
prompt-toolkit==2.0.9
protobuf==3.9.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
pyandoc==0.2.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyCronofy==1.7.1
pycryptodome==3.8.2
PyDispatcher==2.0.5
pyexchange==0.6
Pygments==2.3.1
PyHamcrest==1.9.0
PyInstaller==3.5
PyJWT==1.7.1
pymongo==3.8.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pypandoc==1.4
pyparsing==2.4.0
pytest==4.6.3
pyth==0.6.0
python-dateutil==2.8.0
python-engineio==3.8.2.post1
python-gflags==3.1.2
python-google-calendar-api==0.5
python-pptx==0.6.18
python-socketio==4.2.0
python3-openid==3.1.0
pytz==2018.9
queuelib==1.5.0
Rails==0.0.5
redis==3.3.8
requests==2.21.0
requests-ntlm==1.1.0
requests-oauthlib==1.2.0
rfc3339==6.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
schedule==0.6.0
Scrapy==1.6.0
selenium==3.141.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.8
SpeechRecognition==3.8.1
SQLAlchemy==1.3.2
sqlparse==0.3.0
static3==0.7.0
stringcase==1.2.0
termcolor==1.1.0
textract==1.6.3
traitlets==4.3.2
Twisted==18.9.0
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
w3lib==1.20.0
wcwidth==0.1.7
webargs==5.2.0
WebOb==1.8.5
Werkzeug==0.15.1
whitenoise==4.1.2
xlrd==1.2.0
XlsxWriter==1.1.9
zipp==0.5.1
zope.interface==4.6.0
